In PhotoSection.jsx I added a line to keep the same photo index selected if a new style is selected on the right. If you are on the third image of black joggers and then click the grey joggers style, the image displayed will be the third image of the grey joggers.
In InfoSection, I added functionality for the Add to Cart button. If there is no stock available for a product, the button will be hidden and the dropdown will become inactive and read 'OUT OF STOCK'. I added some functionality for when the button is clicked with no size selected. An error message will pop up telling the user to select a size. Currently working on having the dropdown open as well.